5429: Quick fixes

+++ Web/Edubase.Web.UI/Controllers/TrustController.cs
@@ -162,7 +162,7 @@ namespace Edubase.Web.UI.Controllers
         }
 
 
-        public ActionResult Details(short id)
+        public ActionResult Details(int id)
         {
             using (var dc = new ApplicationDbContext())
             {
